backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=3.0.0
passlib==1.7.4
python-dotenv>=0.19.0
pydantic>=1.8.0
python-multipart>=0.0.5
pyjwt>=2.0.0
email-validator>=1.3.0
bcrypt==4.0.1

# Audio processing / Whisper
whisper>=20230428
ffmpeg-python>=0.2.0

# Real-time communication
python-socketio[asyncio]>=5.0.4

# OpenAI API
openai>=1.32.0
httpx==0.24.0
